% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHexperimenter.R
\name{as.character.IOHexperimenter}
\alias{as.character.IOHexperimenter}
\title{S3 generic as.character function for IOHexperimenter}
\usage{
\method{as.character}{IOHexperimenter}(x, ...)
}
\arguments{
\item{x}{The IOHexperimenter to print}

\item{...}{Arguments for underlying function}
}
\description{
S3 generic as.character function for IOHexperimenter
}
