% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHproblem.R
\name{as.character.IOHproblem}
\alias{as.character.IOHproblem}
\title{S3 generic as.character function for IOHproblem}
\usage{
\method{as.character}{IOHproblem}(x, ...)
}
\arguments{
\item{x}{The IOHproblem to print}

\item{...}{Arguments for underlying function}
}
\description{
S3 generic as.character function for IOHproblem
}
\examples{
exp <- IOHexperimenter()
p <- next_problem(exp)
as.character(p)
}
