% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Benchmarking.R
\name{benchmark_algorithm}
\alias{benchmark_algorithm}
\title{Base procedure for benchmarking a custom algorithm}
\usage{
benchmark_algorithm(user_alg, functions = NULL, instances = NULL,
  dimensions = NULL, data.dir = NULL, algorithm.info = "",
  algorithm.name = "", cdat = FALSE, idat = 0, tdat = 3,
  params.track = NULL, repetitions = 5)
}
\arguments{
\item{user_alg}{Function defining the custom algorithm. Needs to accept one paramter:
an IOHproblem object, which has the following properties:
\itemize{
\item dimension
\item function_id
\item instance
\item fopt (if known)
\item xopt (if known)
}
And the following functions:
\itemize{
\item obj_func
\item target_hit
\item set_parameters
}}

\item{functions}{Which function to test on}

\item{instances}{Which instances to test on}

\item{dimensions}{Which dimensions to test on}

\item{data.dir}{Where the data should be stored (defaults to "./data" when not provided)}

\item{algorithm.info}{Additional information about the algorithm you plan on running}

\item{algorithm.name}{The name of the algorithm you plan on running}

\item{cdat}{Whether or not to generate a .cdat-file}

\item{idat}{Integer}

\item{tdat}{What frequency to use in a .tdat-file}

\item{params.track}{Which parameters to track. Should be a vector of strings, containing no spaces or commas}

\item{repetitions}{How many independent runs of the algorithm to do for each problem}
}
\description{
Base procedure for benchmarking a custom algorithm
}
\examples{
\donttest{
random_search <- function(IOHproblem) {
iter <- 0
fopt <- -Inf
max_iter <- IOHproblem$dimension
while (iter < max_iter && !IOHproblem$target_hit()){
  candidate <- sample(c(0, 1), IOHproblem$dimension, TRUE)
  IOHproblem$set_parameters(0)
  fopt <- max(IOHproblem$obj_func(candidate), fopt)
  iter <- iter+1
}
}

benchmark_algorithm(random_search, params.track = 'Test_param')
}
}
