% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHexperimenter.R
\name{cat.IOHexperimenter}
\alias{cat.IOHexperimenter}
\title{S3 generic cat function for IOHexperimenter}
\usage{
cat.IOHexperimenter(x)
}
\arguments{
\item{x}{The IOHexperimenter to print}
}
\description{
S3 generic cat function for IOHexperimenter
}
