% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHproblem.R
\name{cat.IOHproblem}
\alias{cat.IOHproblem}
\title{S3 generic cat function for IOHproblem}
\usage{
cat.IOHproblem(x)
}
\arguments{
\item{x}{The IOHproblem to print}
}
\description{
S3 generic cat function for IOHproblem
}
