% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHexperimenter.R
\name{get_evaluations}
\alias{get_evaluations}
\title{Get eveluations of the current function / suite?}
\usage{
get_evaluations()
}
\value{
something
}
\description{
Get eveluations of the current function / suite?
}
