% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHexperimenter.R
\name{IOHexperimenter}
\alias{IOHexperimenter}
\title{S3 class 'IOHexperimenter'}
\usage{
IOHexperimenter(dims = c(100, 500, 1000, 2000, 3000),
  functions = seq(23), instances = seq(100), algorithm_info = "",
  algorithm_name = "", data.dir = "./data", cdat = FALSE, idat = 0,
  tdat = 3, param.track = NULL)
}
\arguments{
\item{dims}{Numerical}

\item{functions}{Numerical}

\item{instances}{Numerical Whether the underlying optimization algorithm performs a maximization?}
}
\value{
A S3 object 'DataSet'
}
\description{
S3 class 'IOHexperimenter'
}
