% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms.R
\name{mutate}
\alias{mutate}
\title{Mutation operator for 1+lambda EA}
\usage{
mutate(ind, mutation_rate)
}
\arguments{
\item{ind}{The individual to mutate}

\item{mutation_rate}{The mutation rate}
}
\description{
Mutation operator for 1+lambda EA
}
