% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHexperimenter.R
\name{next_function}
\alias{next_function}
\title{Get the next function of the currently initialized suite}
\usage{
next_function(exp)
}
\arguments{
\item{exp}{Something?}
}
\value{
Something?
}
\description{
Get the next function of the currently initialized suite
}
