% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHproblem.R
\name{next_problem}
\alias{next_problem}
\title{Get the next function of the currently initialized IOHexperimenter object}
\usage{
next_problem(experimenter)
}
\arguments{
\item{experimenter}{The IOHexperimenter object}
}
\value{
An IOHproblem object if available, NULL otherwise
}
\description{
Get the next function of the currently initialized IOHexperimenter object
}
\examples{
exp <- IOHexperimenter()
p <- next_problem(exp)
}
