% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHexperimenter.R
\name{obj_func}
\alias{obj_func}
\title{Evaluate the current function}
\usage{
obj_func(prob, x)
}
\arguments{
\item{prob}{the current IOHproblem}

\item{x}{point to evaluate}
}
\value{
The value of f(x) in the current problem instance
}
\description{
Evaluate the current function
}
