% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms.R
\name{one_plus_lambda_EA}
\alias{one_plus_lambda_EA}
\title{A (1+lambda) EA implementation}
\usage{
one_plus_lambda_EA(IOHproblem, lambda_ = 1, budget = NULL)
}
\arguments{
\item{IOHproblem}{An IOHproblem object}

\item{lambda_}{The size of the offspring}

\item{budget}{How many times the objective function can be evaluated}
}
\description{
A (1+lambda) EA implementation
}
\examples{
\donttest{
one_plus_two_EA <- function(IOHproblem) { one_plus_lambda_EA(IOHproblem, lambda_=2) }

benchmark_algorithm(one_plut_two_EA, params.track = "Mutation rate",
algorithm.name = "one_plus_two_EA",
algorithm.info = "Using one_plus_lambda_EA with specific parameter" )
}
}
