% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHproblem.R
\name{print.IOHproblem}
\alias{print.IOHproblem}
\title{S3 print function for IOHproblem}
\usage{
\method{print}{IOHproblem}(x, ...)
}
\arguments{
\item{x}{The IOHproblem to print}

\item{...}{Arguments for underlying function}
}
\description{
S3 print function for IOHproblem
}
\examples{
exp <- IOHexperimenter()
p <- next_problem(exp)
print(p)
}
