% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms.R
\name{random_local_search}
\alias{random_local_search}
\title{A random local search implementation}
\usage{
random_local_search(IOHproblem, budget = NULL)
}
\arguments{
\item{IOHproblem}{An IOHproblem object}

\item{budget}{How many times the objective function can be evaluated}
}
\description{
This implements a bare-bones random local search algorithm.
This serves as an example of how to implement an optimization algorithm
for use with the IOHexperimenter.
}
