% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms.R
\name{random_search}
\alias{random_search}
\title{A random_search implementation}
\usage{
random_search(IOHproblem)
}
\arguments{
\item{IOHproblem}{An IOHproblem object}
}
\description{
This implements a bare-bones random search algorithm.
This serves as an example of how to implement an optimization algorithm
for use with the IOHexperimenter.
}
