% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms.R
\name{RLS}
\alias{RLS}
\title{Random Local Search (RLS) Algorithm}
\usage{
RLS(dim, obj_func, target = NULL, budget = NULL)
}
\arguments{
\item{dim}{integer, the dimension of the}

\item{obj_func}{the objective function to be maximized}

\item{target}{double, the target value to hit}

\item{budget}{integer, maximal allowable number of function evaluations}
}
\description{
The simplest stochastic optimization algorithm for discrete problems. A randomly 
chosen position in the solution vector is perturbated in each iteration. Only
improvements are accepted after perturbation.
}
