% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHexperimenter.R
\name{summary.IOHexperimenter}
\alias{summary.IOHexperimenter}
\title{S3 generic summary operator for IOHexperimenter}
\usage{
\method{summary}{IOHexperimenter}(object, ...)
}
\arguments{
\item{object}{A IOHexperimenter object}

\item{...}{Arguments passed to other methods}
}
\value{
A summary of the IOHexperimenter object.
}
\description{
S3 generic summary operator for IOHexperimenter
}
\examples{
summary(IOHexperimenter())
}
