% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms.R
\name{two_rate_GA}
\alias{two_rate_GA}
\title{One-Comma-Lambda Genetic Algorithm with 2-rate self-adaptive mutation rate}
\usage{
two_rate_GA(IOHproblem, lambda_ = 2, budget = NULL)
}
\arguments{
\item{IOHproblem}{An IOHproblem object}

\item{lambda_}{integer, the population size > 1}

\item{budget}{How many times the objective function can be evaluated}
}
\description{
A genetic algorithm that controls the mutation rate (strength) using the so-called 
2-rate self-adaptation mechanism: the mutation rate is based on a parameter r. For
each generation, half offspring are generated by mutation rate 2r/dim, and half by 
r/2dim. r that the best offspring has been created with will be inherited by 
probability 3/4, the other by 1/4.
}
