% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHexperimenter.R
\name{benchmark_algorithm}
\alias{benchmark_algorithm}
\title{Base procedure for benchmarking a custom algorithm}
\usage{
benchmark_algorithm(user_alg, functions = NULL, instances = NULL,
  dimensions = NULL, data.dir = NULL, algorithm.info = "",
  algorithm.name = "", cdat = FALSE, idat = 0, tdat = 3,
  params.track = "")
}
\arguments{
\item{user_alg}{Function defining the custom algorithm. Needs to accept two paramters:
The objective function (see \code{obj_func}) and an IOHproblem object, which has the following
properties:
\itemize{
\item dimension
\item function_id
\item instance
\item fopt (if known)
\item xopt (if known)
}}

\item{functions}{Which function to test on}

\item{instances}{Which instances to test on}

\item{dimensions}{Which dimensions to test on}

\item{data.dir}{Where the data should be stored (defaults to "./data" when not provided)}

\item{algorithm.info}{Additional information about the algorithm you plan on running}

\item{algorithm.name}{The name of the algorithm you plan on running}

\item{cdat}{Whether or not to generate a .cdat-file}

\item{idat}{Integer}

\item{tdat}{What frequency to use in a .tdat-file}

\item{params.track}{Which parameters to track}
}
\description{
Base procedure for benchmarking a custom algorithm
}
\examples{
\dontrun{
random_search <- function(objective, problem_info) {
iter <- 0
fopt <- -Inf
max_iter <- problem_info$dimension
while (iter < max_iter && !is_target_hit(problem_info)){
  candidate <- sample(c(0, 1), problem_info$dimension, TRUE)
  fopt <- max(objective(problem_info, candidate), fopt)
  iter <- iter+1
}
print(fopt)
}

benchmark_algorithm(random_search)
}
}
