% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHexperimenter.R
\name{IOHexperimenter}
\alias{IOHexperimenter}
\title{S3 class 'IOHexperimenter'}
\usage{
IOHexperimenter(suite = "PBO", dims = NULL, functions = NULL,
  instances = NULL, algorithm.info = " ", algorithm.name = " ",
  data.dir = "./data", param.track = NULL)
}
\arguments{
\item{suite}{Which suite to use. Available: 'PBO', 'BBOB'}

\item{dims}{Numerical}

\item{functions}{Numerical}

\item{instances}{Numerical Whether the underlying optimization algorithm performs a maximization?}

\item{algorithm.info}{Additional information about the algorithm you plan on running}

\item{algorithm.name}{The name of the algorithm you plan on running}

\item{data.dir}{Where the data should be stored}

\item{param.track}{Which parameters to track. Should be a vector of strings, containing no spaces or commas}

\item{cdat}{Whether or not to generate a .cdat-file}

\item{idat}{Integer}

\item{tdat}{What frequency to use in a .tdat-file}
}
\value{
A S3 object 'DataSet'
}
\description{
S3 class 'IOHexperimenter'
}
\examples{
exp <- IOHexperimenter()
}
