% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHexperimenter.R
\name{IOHexperimenter}
\alias{IOHexperimenter}
\title{S3 class 'IOHexperimenter'}
\usage{
IOHexperimenter(dims = c(100, 500, 1000, 2000, 3000),
  functions = seq(23), instances = seq(100), algorithm_info = "",
  algorithm_name = "", data.dir = "./data", cdat = FALSE, idat = 0,
  tdat = 3, param.track = NULL)
}
\arguments{
\item{dims}{Numerical}

\item{functions}{Numerical}

\item{instances}{Numerical Whether the underlying optimization algorithm performs a maximization?}

\item{algorithm_info}{Additional information about the algorithm you plan on running}

\item{algorithm_name}{The name of the algorithm you plan on running}

\item{data.dir}{Where the data should be stored}

\item{cdat}{Whether or not to generate a .cdat-file}

\item{idat}{Integer}

\item{tdat}{What frequency to use in a .tdat-file}

\item{param.track}{Which parameters to track. Should be a vector of strings, containing no spaces or commas}
}
\value{
A S3 object 'DataSet'
}
\description{
S3 class 'IOHexperimenter'
}
