% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHexperimenter.R
\name{IOHexperimenter}
\alias{IOHexperimenter}
\title{S3 class 'IOHexperimenter'}
\usage{
IOHexperimenter(suite = "PBO", dims = NULL, functions = NULL,
  instances = NULL, algorithm.info = " ", algorithm.name = " ",
  data.dir = NULL, param.track = NULL)
}
\arguments{
\item{suite}{Which suite to use. Available: 'PBO', 'BBOB'}

\item{dims}{Numerical Which dimensions to use}

\item{functions}{Numerical Which functions from the selected suite to use}

\item{instances}{Numerical Which problem instances to use}

\item{algorithm.info}{Additional information about the algorithm you plan on running}

\item{algorithm.name}{The name of the algorithm you plan on running}

\item{data.dir}{Where the data should be stored. Defaults to NULL, meaning no data is stored.}

\item{param.track}{Which parameters to track. Should be a vector of strings, containing no spaces or commas}
}
\value{
A S3 object 'DataSet'
}
\description{
S3 class 'IOHexperimenter'
}
\examples{
exp <- IOHexperimenter()
}
