% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHexperimenter.R
\name{next_problem}
\alias{next_problem}
\title{Get the next function of the currently initialized suite}
\usage{
next_problem(experimenter)
}
\arguments{
\item{experimenter}{The IOHexperimenter object}
}
\value{
An IOHproblem object
}
\description{
Get the next function of the currently initialized suite
}
