% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHexperimenter.R
\name{obj_func}
\alias{obj_func}
\title{Evaluate the current function}
\usage{
obj_func(exp, x)
}
\arguments{
\item{exp}{the current IOHexperimenter}

\item{x}{point to evaluate}
}
\description{
Evaluate the current function
}
