% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHexperimenter.R
\name{print.IOHexperimenter}
\alias{print.IOHexperimenter}
\title{S3 print function for IOHexperimenter}
\usage{
\method{print}{IOHexperimenter}(x, ...)
}
\arguments{
\item{x}{The IOHexperimenter to print}

\item{...}{Arguments for underlying function}
}
\description{
S3 print function for IOHexperimenter
}
\examples{
print(IOHexperimenter())
}
