% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms.R
\name{random_local_search}
\alias{random_local_search}
\title{Random Local Search (RLS) Algorithm}
\usage{
random_local_search(IOHproblem, budget = NULL)
}
\arguments{
\item{IOHproblem}{An IOHproblem object}

\item{budget}{integer, maximal allowable number of function evaluations}
}
\description{
The simplest stochastic optimization algorithm for discrete problems. A randomly
chosen position in the solution vector is perturbated in each iteration. Only
improvements are accepted after perturbation.
}
