% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms.R
\name{IOH_random_local_search}
\alias{IOH_random_local_search}
\alias{random_local_search}
\title{IOHexperimenter-based wrapper}
\usage{
IOH_random_local_search(IOHproblem, budget = NULL)

random_local_search(dimension, obj_func, target_hit = function() {    
  FALSE }, budget = NULL)
}
\arguments{
\item{IOHproblem}{An IOHproblem object}

\item{budget}{integer, maximal allowable number of function evaluations}

\item{dimension}{Dimension of search space}

\item{obj_func}{The evaluation function}

\item{target_hit}{Optional, function which enables early stopping if a target value is reached}
}
\value{
A list containing the location (xopt) and value (fopt) of the optimal value found
}
\description{
For easier use with the IOHexperimenter

The simplest stochastic optimization algorithm for discrete problems. A randomly
chosen position in the solution vector is perturbated in each iteration. Only
improvements are accepted after perturbation.
}
\examples{
\donttest{
benchmark_algorithm(IOH_random_local_search, data.dir = NULL)
}
}
