% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms.R
\name{IOH_random_search}
\alias{IOH_random_search}
\alias{random_search}
\title{IOHexperimenter-based wrapper}
\usage{
IOH_random_search(IOHproblem, budget = NULL)

random_search(dim, obj_func, target = NULL, budget = NULL)
}
\arguments{
\item{IOHproblem}{An IOHproblem object}

\item{budget}{integer, maximal allowable number of function evaluations}

\item{dim}{Dimension of search space}

\item{obj_func}{The evaluation function}

\item{target}{Optional, enables early stopping if this value is reached}
}
\description{
For easier use with the IOHexperimenter

Random walk in \eqn{{0, 1}^d} space
}
\examples{
benchmark_algorithm(IOH_random_search)
}
