% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms.R
\name{random_search}
\alias{random_search}
\title{Random Search}
\usage{
random_search(IOHproblem, budget = NULL)
}
\arguments{
\item{IOHproblem}{An IOHproblem object}

\item{budget}{integer, maximal allowable number of function evaluations}
}
\description{
Random walk in \eqn{{0, 1}^d} space
}
