% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOHproblem.R
\name{reset_problem}
\alias{reset_problem}
\title{Reset the IOHproblem}
\usage{
reset_problem(problem)
}
\arguments{
\item{problem}{The IOHproblem object}
}
\value{
An IOHproblem object
}
\description{
Reset the IOHproblem
}
